Create TheGame.java
This code is based on the week4 version of MOOC "Begin programming: Begin programming: build your first mobile game" of the University of Reading. It contains two modifications by A. Andreatos: 
a) It displays a smiley at the top of the screen (from week 4). Smiley moves back and forth horizontally. b) It displays a horizontal bar to act like a racket; when the ball hits the bar, it changes direction.