Create TheGame-function.java
TheGame-function.java, to replace file "TheGame.java".
Answer to activity "6.5 Creating a function".
Contains a function placeMiddleBottom()  that repositions the paddle into the middle-bottom of the screen each time we score. 

The function is: private void placeMiddleBottom() at line 172. 
It also contains the ‘updateBallCollusion’ function as defined on p.16 of "6-2_Guide_to_functions.pdf" (Line 177). 
By A. Andreatos, 31 march 2014.